+++ python/docs/tools.md
@@ -369,7 +369,7 @@ from mcp.client.stdio import stdio_client
 from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.agents.react import ReActAgent
 from beeai_framework.memory import UnconstrainedMemory
-from beeai_framework.tools import MCPTool
+from beeai_framework.tools.mcp import MCPTool
 
 load_dotenv()
