+++ python/examples/workflows/multi_agents_simple.py
@@ -7,7 +7,7 @@ from beeai_framework.emitter import EmitterOptions
 from beeai_framework.errors import FrameworkError
 from beeai_framework.tools.search import WikipediaTool
 from beeai_framework.tools.weather import OpenMeteoTool
-from beeai_framework.workflows import AgentWorkflow, AgentWorkflowInput
+from beeai_framework.workflows.agent import AgentWorkflow, AgentWorkflowInput
 from examples.helpers.io import ConsoleReader
