+++ python/examples/workflows/travel_advisor.py
@@ -6,7 +6,7 @@ from beeai_framework.backend import ChatModel
 from beeai_framework.errors import FrameworkError
 from beeai_framework.tools.search import DuckDuckGoSearchTool
 from beeai_framework.tools.weather import OpenMeteoTool
-from beeai_framework.workflows import AgentWorkflow, AgentWorkflowInput
+from beeai_framework.workflows.agent import AgentWorkflow, AgentWorkflowInput
 
 
 async def main() -> None:
