+++ python/tests/workflows/test_multi_agents.py
@@ -17,7 +17,7 @@ import pytest
 from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.backend import UserMessage
 from beeai_framework.memory import UnconstrainedMemory
-from beeai_framework.workflows import AgentWorkflow
+from beeai_framework.workflows.agent import AgentWorkflow
 
 """
 E2E Tests
